Consolidated popover architecture
